Final State BugFix
This is a bugfix pertaining to the return of Irand and Genenerator in
Final state due to increment of i. This has been fixed and verification
code has been added.

Also, the gitignore file has been edited to exclude asv files.

This code also has The state variable WeightDeriv inserted. This
variable is specific to the Working Mem simulations so this is
unfortunate but It doesnt affect performance, so meh.