Memory BugFixes plus Memory Manage
BugFix is pertaining to trim() function where some of the if
conditioning was just plain wrong. since that function was never really
called upon in any non-trivial circumstance, it was unnoticed up until
now.

Also changed mxFree()'s to resize(0); trim(); statement pairs. This is
because the trim() function ensures destruction of all the elements when
de/reallocating array

Introduced new class MexCounter with private member MemUsageCounter and
Public Member MemUsageLimit. MemUsageCounter is accessible to all
MexVector and MexMatrix classes. The methods in these classes are
programmed to check the amount of total memory consumed and required and
accordingly take decisions or throw ExOps::MEM_FULL_... Exception.

The Exception Processing Code has been written in the MexFunction. The
exception has also been thrown back to the main() function in the case
of the EXE interface.