Added STDP + MANY Changes
The List of changes is

1. Added vectors ExpVect in order to facilitate the exponential required
for STDP.
2. Changed all the ints in IntVars to size_t.
3. Added NExc, MExc in order to perform processing differently for STDP
of Inh and Exc Synapses
4. Edited the convention for the PreSyn....End/Beg and PostSyn...End/Beg
so that the end and beg are continuous and never -1 (as before). The
corresponding code in SpikeStore has been verified to be correct. This
was done to calculate MExc from NExc
5. Added Causal and Anti Causal STDP Rule
6. Added the once a second update rule.
7. This has the important issue that the static vectors cause shitty
behaviour with MEX files during destruction.
Not clear yet why. So, I will edit in main branch and merge with this.
8. Changed ConvertStatetoInitialCond.m
9. Edited ConsistencyTest.m for this pearticular version