PreSynNeuronSection..... BugFix
This is a bugfix for the issue where the zeroth element of the Section Arrays was not initialized to 0. This is potentially a problem where the first neuron does not have any fanouts thereby leaving that neuron's array indices at -1 which is contrary to the rule that the indices must be continuous.