Temporary Untested Commit
This is the first (untested) complete commit for the Polychronous group
finding algorithm. It is used to detect groups generated by neuron
triplets. This needs to undergo testing for which the needs to be
modified to find PNGs generated by duplets. When that works itl be
checked in along with this commit