BugFix 1-No Jitter Edition
This is the Bugfix having been tested for bugs using the Duplet version of the code on the toy network. Currently, it is not coded to accept 1 ms Jitter. This is planned for the next edition.

In this all the matlab code relating to Preparing the simulation, Reading, Processing and Displaying PNGs have been written and tested.