Refactoring Array Resetting
Pushed array resetting into a function different from AnalyseGroups so that Analyse groups could be optionally disabled easily.

The function is (ResetIntermediateVars(SimVars))

The DisplayPNG.m is changed. in order to make visualization easier, the Random Jitterintroduced in the Y-coordinates is maintaind constant across different points with the same Y-coordinates.

Test.m is changed to include code for mex function call as well as call the changed DisplayPNG Function.