Added 1ms Jitter Compliance
The code here allows synapses that arrived one ms earlier to contribute to the spike at the current time instant. In the event of a spike, the neuron is not allowed to spike in the next ms.

The ConvertPNGtoRelative.m is changed accordingly to search for neurons spiking in the 2ms interval instead of one ms interval.